refactor: implemement ContentFilters.azure() and deprecate buildAzureContentFilter()
