+++ packages/orchestration/README.md
@@ -245,12 +245,9 @@ Use the orchestration client with filtering to restrict content that is passed t
 This feature allows filtering both the [input](https://help.sap.com/docs/sap-ai-core/sap-ai-core-service-guide/consume-orchestration#content-filtering-on-input) and [output](https://help.sap.com/docs/sap-ai-core/sap-ai-core-service-guide/consume-orchestration#content-filtering-on-input) of a model based on content safety criteria.
 
 ```ts
-import {
-  OrchestrationClient,
-  buildAzureContentFilter
-} from '@sap-ai-sdk/orchestration';
+import { OrchestrationClient, ContentFilters } from '@sap-ai-sdk/orchestration';
 
-const filter = buildAzureContentFilter({ Hate: 2, Violence: 4 });
+const filter = ContentFilters.azure({ Hate: 2, Violence: 4 });
 const orchestrationClient = new OrchestrationClient({
   llm: {
     model_name: 'gpt-4o',
@@ -260,8 +257,12 @@ const orchestrationClient = new OrchestrationClient({
     template: [{ role: 'user', content: '{{?input}}' }]
   },
   filtering: {
-    input: filter,
-    output: filter
+    input: {
+      filters: [filter]
+    },
+    output: {
+      filters: [filter]
+    }
   }
 });
 
@@ -288,7 +289,11 @@ Both `chatCompletion()` and `getContent()` methods can throw errors.
 
 Therefore, handle errors appropriately to ensure meaningful feedback for both types of errors.
 
-`buildAzureContentFilter()` is a convenience function that creates an Azure content filter configuration based on the provided inputs.
+`ContentFilters` contains helper functions to create content filters.
+
+#### Azure Content Filter
+
+Use `ContentFilters.azure()` to build an Azure content filter.
 The Azure content filter supports four categories: `Hate`, `Violence`, `Sexual`, and `SelfHarm`.
 Each category can be configured with severity levels of 0, 2, 4, or 6.
