+++ packages/orchestration/src/internal.ts
@@ -1,4 +1,4 @@
 export * from './orchestration-client.js';
-export * from './orchestration-utils.js';
+export * from './orchestration-filtering.js';
 export * from './orchestration-types.js';
 export * from './orchestration-response.js';
