--- packages/orchestration/src/orchestration-utils.ts
@@ -1,26 +1,0 @@
-import type {
-  AzureContentSafety,
-  InputFilteringConfig,
-  OutputFilteringConfig
-} from './client/api/schema/index.js';
-
-/**
- * Convenience function to create Azure content filters.
- * @param filter - Filtering configuration for Azure filter. If skipped, the default Azure content filter configuration is used.
- * @returns An object with the Azure filtering configuration.
- */
-export function buildAzureContentFilter(
-  filter?: AzureContentSafety
-): InputFilteringConfig | OutputFilteringConfig {
-  if (filter && !Object.keys(filter).length) {
-    throw new Error('Filter property cannot be an empty object');
-  }
-  return {
-    filters: [
-      {
-        type: 'azure_content_safety',
-        ...(filter && { config: filter })
-      }
-    ]
-  };
-}
