+++ sample-code/src/orchestration.ts
@@ -1,8 +1,5 @@
 import { readFile } from 'node:fs/promises';
-import {
-  OrchestrationClient,
-  buildAzureContentFilter
-} from '@sap-ai-sdk/orchestration';
+import { ContentFilters, OrchestrationClient } from '@sap-ai-sdk/orchestration';
 import { createLogger } from '@sap-cloud-sdk/util';
 import type {
   LlmModuleConfig,
@@ -72,13 +69,15 @@ const templating = { template: [{ role: 'user', content: '{{?input}}' }] };
 export async function orchestrationInputFiltering(): Promise<void> {
   // create a filter with minimal thresholds for hate and violence
   // lower numbers mean more strict filtering
-  const filter = buildAzureContentFilter({ Hate: 0, Violence: 0 });
+  const azureContentFilter = ContentFilters.azure({ Hate: 0, Violence: 0 });
   const orchestrationClient = new OrchestrationClient({
     llm,
     templating,
     // configure the filter to be applied for both input and output
     filtering: {
-      input: filter
+      input: {
+        filters: [azureContentFilter]
+      }
     }
   });
 
@@ -104,12 +103,14 @@ export async function orchestrationInputFiltering(): Promise<void> {
 export async function orchestrationOutputFiltering(): Promise<OrchestrationResponse> {
   // output filters are build in the same way as input filters
   // set the thresholds to the minimum to maximize the chance the LLM output will be filtered
-  const filter = buildAzureContentFilter({ Hate: 0, Violence: 0 });
+  const azureContentFilter = ContentFilters.azure({ Hate: 0, Violence: 0 });
   const orchestrationClient = new OrchestrationClient({
     llm,
     templating,
     filtering: {
-      output: filter
+      output: {
+        filters: [azureContentFilter]
+      }
     }
   });
   /**
